fix: 调整一些代码
